cdn/requirements.txt
Flask
Flask-Uploads
Pillow